Refactor etag handling to separate file
