+++ lib/file.js
@@ -2,12 +2,11 @@
 
 var Fs = require('fs');
 var Path = require('path');
-var Crypto = require('crypto');
 var Ammo = require('ammo');
 var Boom = require('boom');
 var Hoek = require('hoek');
 var Joi = require('joi');
-var LruCache = require('lru-cache');
+var Etag = require('./etag');
 
 
 // Declare internals
@@ -93,7 +92,7 @@ internals.prepare = function (response, callback) {
             response.header('content-disposition', response.source.settings.mode + '; filename=' + encodeURIComponent(fileName));
         }
 
-        internals.addEtag(response, stat, function (err) {
+        Etag.apply(response, stat, function (err) {
 
             if (err) {
                 return callback(err);
@@ -130,80 +129,6 @@ internals.marshal = function (response, next) {
 
         return internals.openStream(response, gzFile, next);
     });
-};
-
-
-internals.addHashedEtag = function (response, stat, callback) {
-
-    var etags = response.request.server.plugins.inert._etags;
-    if (!etags) {
-        return callback();
-    }
-
-    // Use stat info for an LRU cache key.
-
-    var path = response.source.path;
-    var cachekey = [path, stat.ino, stat.size, stat.mtime.getTime()].join('-');
-
-    // The etag hashes the file contents in order to be consistent across distributed deployments
-
-    var cachedEtag = etags.get(cachekey);
-    if (cachedEtag) {
-        response.etag(cachedEtag, { vary: true });
-        return callback();
-    }
-
-    // Perform the hashing
-
-    var hash = Crypto.createHash('sha1');
-    hash.setEncoding('hex');
-
-    var fileStream = Fs.createReadStream(path, { fd: response.source.fd, autoClose: false });
-    fileStream.pipe(hash);
-
-    var done = function (err) {
-
-        if (err) {
-            return callback(Boom.wrap(err, null, 'Failed to hash file'));
-        }
-
-        var etag = hash.read();
-        etags.set(cachekey, etag);
-
-        response.etag(etag, { vary: true });
-
-        return callback();
-    };
-
-    done = Hoek.once(done);
-
-    fileStream.on('end', done);
-    fileStream.on('error', done);
-};
-
-
-internals.addSimpleEtag = function (response, stat, callback) {
-
-    var size = stat.size.toString(16);
-    var mtime = stat.mtime.getTime().toString(16);
-
-    response.etag(size + '-' + mtime, { vary: true });
-
-    return callback();
-};
-
-
-internals.addEtag = function (response, stat, callback) {
-
-    if (response.source.settings.etagMethod === false) {
-        return callback();
-    }
-
-    if (response.source.settings.etagMethod === 'simple') {
-        return internals.addSimpleEtag(response, stat, callback);
-    }
-
-    return internals.addHashedEtag(response, stat, callback);
 };
 
 
@@ -312,6 +237,3 @@ internals.close = function (response) {
         response.source.fd = null;
     }
 };
-
-
-exports.Etags = LruCache;
