+++ lib/index.js
@@ -1,6 +1,7 @@
 // Load modules
 
 var Directory = require('./directory');
+var Etag = require('./etag');
 var File = require('./file');
 var Hoek = require('hoek');
 
@@ -18,7 +19,7 @@ exports.register = function (server, options, next) {
 
     var settings = Hoek.applyToDefaults(internals.defaults, options);
 
-    server.expose('_etags', settings.etagsCacheMaxSize ? new File.Etags(settings.etagsCacheMaxSize) : null);
+    server.expose('_etags', settings.etagsCacheMaxSize ? new Etag.Cache(settings.etagsCacheMaxSize) : null);
 
     server.handler('file', File.handler);
     server.handler('directory', Directory.handler);
